dry & kiss
remove repetitive code with a simpler logic
